refactor(check-collections): relocate CheckCollectionResult to base + dataclass-ify
Move ``CheckCollectionResult`` from ``contracts/contract_verification.py``
into ``check_collections/base.py`` (its spec home) and apply ``@dataclass``
to both it and ``CheckCollectionSessionResult``.

``contract_verification.py`` re-imports the class at module bottom — that
order matters because ``base.py`` needs ``Contract`` / ``DataSource`` /
``Measurement`` / ``CheckResult`` / ``PostProcessingStage`` /
``ScanTokenUsage`` / ``YamlFileContentInfo`` / ``ContractVerificationStatus``
defined above it. ``ContractVerificationResult`` stays as a thin
``@dataclass`` subclass preserving the historical name, so existing
keyword-argument call sites in tests and ``soda_cloud.py`` continue to work
unchanged.
